requirements.txt
flask
configargparser
python_dateutil == 2.6.0
setuptools >= 21.0.0
pynacl
openapi-core==0.19.0a1
psutil
configargparse
ptw
pynacl
requests
kivy[base]
kivy-garden.qrcode
kivy-garden
proxy.py
cx_freeze
